Allowing decimal zoom values
by request by Zenol

+++ viewer_view.php
@@ -150,7 +150,7 @@ $viewer->load_ts();
 
 if(array_key_exists('zoom',$_REQUEST)) {
 	assert(is_numeric($_REQUEST['zoom'])) or die('bad zoom value');
-	$viewer->zoom=intval($_REQUEST['zoom']);
+	$viewer->zoom=floatval($_REQUEST['zoom']);
 	assert($viewer->zoom>=0.1 && $viewer->zoom<=10) or die('bad zoom range');
 }
